% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{barPlotGadget}
\alias{barPlotGadget}
\title{Custom Bar Plot Gadget}
\usage{
barPlotGadget(data_bar)
}
\arguments{
\item{data_bar}{A data frame with columns 'category' and 'value'.}
}
\value{
A Shiny application for custom bar plot.
}
\description{
This Shiny gadget allows users to create a custom bar plot using ggplot2.
}
