% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphFunctions.R
\name{customBarPlot}
\alias{customBarPlot}
\title{Custom Bar Plot}
\usage{
customBarPlot(x, y, fill = "darkgreen")
}
\arguments{
\item{x}{Factor or character vector for the x-axis, representing categories.}

\item{y}{Numeric vector for the y-axis, representing the values of each category.}

\item{fill}{Character, fill color of the bars in the bar plot.}
}
\value{
A ggplot2 bar plot object.
}
\description{
This function creates a customized bar plot using ggplot2. It is designed to visualize the distribution of a numeric variable across different categories.
}
\examples{
# Sample Data
data_bar <- data.frame(category = letters[1:5], value = c(3, 7, 2, 5, 9))

# Create a custom bar plot
customBarPlot(x = data_bar$category, y = data_bar$value, fill = "darkgreen")

}
