% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphFunctions.R
\name{customScatterPlot}
\alias{customScatterPlot}
\title{Custom Scatter Plot}
\usage{
customScatterPlot(x, y, color = "steelblue", size = 3)
}
\arguments{
\item{x}{Numeric vector for the x-axis, representing one variable.}

\item{y}{Numeric vector for the y-axis, representing another variable.}

\item{color}{Character, color of the points on the scatter plot.}

\item{size}{Numeric, size of the points on the scatter plot.}
}
\value{
A ggplot2 scatter plot object.
}
\description{
This function creates a customized scatter plot using ggplot2. It allows you to visualize the relationship between two numeric variables.
}
\examples{
# Sample Data
set.seed(123)
data <- data.frame(x = rnorm(100), y = rnorm(100))

# Create a custom scatter plot
customScatterPlot(x = data$x, y = data$y, color = "steelblue", size = 4)

}
