% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{linePlotGadget}
\alias{linePlotGadget}
\title{Custom Line Plot Gadget}
\usage{
linePlotGadget(data_line)
}
\arguments{
\item{data_line}{A data frame with numeric columns 'x' and 'y'.}
}
\value{
A Shiny application for custom line plot.
}
\description{
This Shiny gadget allows users to create a custom line plot using ggplot2.
}
