% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{scatterPlotGadget}
\alias{scatterPlotGadget}
\title{Custom Scatter Plot Gadget}
\usage{
scatterPlotGadget(data)
}
\arguments{
\item{data}{A data frame with numeric columns 'x' and 'y'.}
}
\value{
A Shiny application for custom scatter plot.
}
\description{
This Shiny gadget allows users to create a custom scatter plot using ggplot2.
}
